backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
requests==2.31.0
beautifulsoup4==4.12.3
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
flask>=3.0.0
flask-cors>=4.0.0
